% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cbindBM}
\alias{cbindBM}
\title{cbind functionality for class "big.matrix"}
\usage{
cbindBM(x, y, binding = "right", cols.x = NULL, cols.y = NULL, z = NULL,
  type = NULL, separated = NULL, backingfile = NULL, backingpath = NULL,
  descriptorfile = NULL, binarydescriptor = FALSE, shared = TRUE)
}
\arguments{
\item{x}{A \code{"big.matrix"}}

\item{y}{Columns to append, accepts \code{"big.matrix"} and \code{"matrix"} in addition to
\code{"numeric"} and \code{"integer"} (e.g. "2.5", "1L")}

\item{binding}{Binding location, "right" or "left"}

\item{cols.x}{Possible subset of columns from the x object; could be numeric, named, or logical}

\item{cols.y}{Possible subset of columns from the y object if a matrix type object otherwise ignored;
could be numeric, named, or logical}

\item{z}{Optional destinitation object (matrix or big.matrix); if not specified, a big.matrix will be created}

\item{type}{preferably specified (e.g. "integer", "double", etc.)}

\item{separated}{use separated column organization of the data instead of column-major organization;
use with caution if the number of columns is large.}

\item{backingfile}{the root name for the file(s) for the cache of x.}

\item{backingpath}{}

\item{descriptorfile}{}

\item{binarydescriptor}{}

\item{shared}{}
}
\value{
a \code{"big.matrix"}
}
\description{
This is used to bind columns to big.matrix objects, with the new copy optionally filebacked
}
\details{
This is simply a wrapper for \code{"big.matrix"} objects whereby a new \code{"big.matrix"} is
initialized and subsequently filled with the submitted arguments (optionally filtered).
}

