% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cbindBMIP}
\alias{cbindBMIP}
\title{cbind 'in-place' for class "big.matrix"}
\usage{
cbindBMIP(x, y, cols.y = NULL)
}
\arguments{
\item{x}{A \code{"big.matrix"}}

\item{y}{Columns to append, accepts \code{"big.matrix"} and \code{"matrix"} in addition to
\code{"numeric"} and \code{"integer"} (e.g. "2.5", "1L")}

\item{cols.y}{Optionally specify columns to bind from 'y' if 'y' is a \code{"matrix"} or
\code{"big.matrix"}.}
}
\value{
The original \code{"big.matrix"} with the added columns
}
\description{
This is used to bind columns to an existing big.matrix object.
WARNING!!! This modifies the original big.matrix object and therefore must be
either recreated or subset to regain the original structure.  Use only if you
intend to only use the original matrix in the new form.
}
\examples{
# Create big.matrix
A <- matrix(seq(12), 3, 4)
BM <- as.big.matrix(A)
C <- as.big.matrix(A)
head(BM)
# Call in-place cbind
# cbind only 1st and 3rd column
cbindBMIP(BM, C, c(1,3))
head(BM)
}

