% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expBMIP}
\alias{expBMIP}
\title{Provide in-place exponential function for class "big.matrix"}
\usage{
expBMIP(x)
}
\arguments{
\item{x}{A \code{"big.matrix"} which must be of type "double" to allow decimals}
}
\value{
a \code{"big.matrix"}
}
\description{
This is used to take each element of a \code{"big.matrix"} to a specified power.
WARNING!!! This replaces the original matrix.
}
\details{
This function takes each element of a \code{"big.matrix"} to its' respective exponential.
}

