% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iptBM}
\alias{iptBM}
\title{'Big.Matrix' In-Place-Transpose}
\usage{
iptBM(x, direction)
}
\arguments{
\item{x}{A \code{"big.matrix"}}

\item{direction}{Specify which direction of transpose}
}
\value{
Nothing, the \code{"big.matrix"} object has been modified.
}
\description{
This is used to apply and in-place transpose of a \code{"big.matrix"}.  It modifies the
original 'big.matrix' to its' corresponding tranpose.  WARNING!!! This changes your big.matrix object
and therefore the original structure is not accessible at the same time.  You can always convert back
to the original structure with this same function.
}
\examples{
# Create big.matrix
BM <- as.big.matrix(matrix(seq(12), 3, 4))
head(BM)
# Call in-place tranpose
iptBM(BM, direction="r2c")
head(BM)
}

