% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logBM}
\alias{logBM}
\title{Provide natural logarithm function for class "big.matrix"}
\usage{
logBM(x, cols.x = NULL, z = NULL, type = "double", separated = NULL,
  backingfile = NULL, backingpath = NULL, descriptorfile = NULL,
  binarydescriptor = FALSE, shared = TRUE)
}
\arguments{
\item{x}{A \code{"big.matrix"}}

\item{cols.x}{Possible subset of columns from the x object; could be numeric, named, or logical}

\item{z}{Optional destinitation object (matrix or big.matrix); if not specified, a big.matrix will be created}

\item{type}{default: "double"}

\item{separated}{use separated column organization of the data instead of column-major organization;
use with caution if the number of columns is large.}

\item{backingfile}{the root name for the file(s) for the cache of x.}

\item{backingpath}{the path to the directory containing the file backing cache}

\item{descriptorfile}{the name of the file to hold the backingfile description, for subsequent use with
\code{\link[bigmemory]{attach.big.matrix}}; if NULL, the backingfile is used as the root part of the descriptor file name.
The descriptor file is placed in the same directory as the backing files.}

\item{binarydescriptor}{the flag to specify if the binary RDS format should be used for the backingfile description,
for subsequent use with \code{\link[bigmemory]{attach.big.matrix}}; if NULL of FALSE, the dput() file format is used.}

\item{shared}{TRUE by default, and always TRUE if the big.matrix is file-backed. For a non-filebacked big.matrix,
shared=FALSE uses non-shared memory, which can be more stable for large (say, >50% of RAM) objects. Shared memory
allocation can sometimes fail in such cases due to exhausted shared-memory resources in the system.}
}
\value{
a \code{"big.matrix"}
}
\description{
This is used to take each element of a \code{"big.matrix"} and return the natural log with the
option to have the new copy filebacked.
}
\details{
This generates a new \code{"big.matrix"} object whereby each element has been replaced with its' natural log
of \code{"value"}.  This function should only need to be used if the new matrix must be filebacked as the
NULL call is applied in the \code{\link[base]{log}} method.
}

