% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logBMIP}
\alias{logBMIP}
\title{Provide in-place natural log function for class "big.matrix"}
\usage{
logBMIP(x, base = exp(1))
}
\arguments{
\item{x}{A \code{"big.matrix"} which must be of type "double" to allow decimals}

\item{base}{The base for the logarithm. Default = \code{"exp(1)"}}
}
\value{
a \code{"big.matrix"}
}
\description{
This is used to take each element of a \code{"big.matrix"} to a specified power.
WARNING!!! This replaces the original matrix.
}
\details{
This function takes each element of a \code{"big.matrix"} to replaes with its' respective base logarithm.
}
\seealso{
\code{\link[base]{log10}} for generic common log10 call, \code{\link[bigmemoryExt]{log10BM}} for more control
of new matrix, and \code{\link[bigmemoryExt]{logBaseBM}} for different logartihm bases.
}

