% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rbindBM}
\alias{rbindBM}
\title{rbind functionality for class "big.matrix"}
\usage{
rbindBM(x, y, binding = "right", rows.x = NULL, rows.y = NULL, z = NULL,
  type = NULL, separated = NULL, backingfile = NULL, backingpath = NULL,
  descriptorfile = NULL, binarydescriptor = FALSE, shared = TRUE)
}
\arguments{
\item{x}{A \code{"big.matrix"}}

\item{y}{Columns to append, accepts \code{"big.matrix"} and \code{"matrix"} in addition to
\code{"numeric"} and \code{"integer"} (e.g. "2.5", "1L")}

\item{binding}{Binding location, "right" or "left"}

\item{cols.x}{Possible subset of rows from the x object; could be numeric, named, or logical}

\item{cols.y}{Possible subset of rows from the y object if a matrix type object otherwise ignored;
could be numeric, named, or logical}

\item{z}{Optional destination object (matrix or big.matrix); if not specified, a big.matrix will be created}

\item{type}{preferably specified (e.g. "integer", "double", etc.)}

\item{separated}{use separated column organization of the data instead of column-major organization;
use with caution if the number of columns is large.}

\item{backingfile}{the root name for the file(s) for the cache of x.}

\item{backingpath}{the path to the directory containing the file backing cache}

\item{descriptorfile}{the name of the file to hold the backingfile description, for subsequent use with
\code{\link[bigmemory]{attach.big.matrix}}; if NULL, the backingfile is used as the root part of the descriptor file name.
The descriptor file is placed in the same directory as the backing files.}

\item{binarydescriptor}{the flag to specify if the binary RDS format should be used for the backingfile description,
for subsequent use with \code{\link[bigmemory]{attach.big.matrix}}; if NULL of FALSE, the dput() file format is used.}

\item{shared}{TRUE by default, and always TRUE if the big.matrix is file-backed. For a non-filebacked big.matrix,
shared=FALSE uses non-shared memory, which can be more stable for large (say, >50% of RAM) objects. Shared memory
allocation can sometimes fail in such cases due to exhausted shared-memory resources in the system.}
}
\value{
a \code{"big.matrix"}
}
\description{
This is used to bind columns to big.matrix objects, with the new copy optionally filebacked
}
\details{
This is simply a wrapper for \code{"big.matrix"} objects whereby a new \code{"big.matrix"} is
initialized and subsequently filled with the submitted arguments (optionally filtered).
}
