% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sinhBMIP}
\alias{sinhBMIP}
\title{Provide in-place hyperbolic sine function for class "big.matrix"}
\usage{
sinhBMIP(x)
}
\arguments{
\item{x}{A \code{"big.matrix"} which must be of type "double" to allow decimals}
}
\value{
a \code{"big.matrix"}
}
\description{
This is used to take each element of a \code{"big.matrix"} to its' hyperbolic sine.
WARNING!!! This replaces the original matrix.
}
\details{
This function takes each element of a \code{"big.matrix"} to its' hyperbolic sine.
}
\seealso{
\code{\link[bigmemoryExt]{sinhBM}} for separate big.matrix
}

