% Generated by roxygen2 (4.0.2): do not edit by hand
\name{subtractIntBM}
\alias{subtractIntBM}
\title{Subtract a "big.matrix" from an integer}
\usage{
subtractIntBM(x, value, cols = NULL, rows = NULL, y = NULL, type = NULL,
  separated = NULL, backingfile = NULL, backingpath = NULL,
  descriptorfile = NULL, binarydescriptor = FALSE, shared = TRUE)
}
\arguments{
\item{x}{A \code{"big.matrix"}}

\item{value}{A numeric value (e.g. "2.5", "1L")}

\item{cols}{Possible subset of columns for the transpose; could be numeric, named, or logical}

\item{rows}{Possible subset of rows for the transpose; could be numeric, named, or logical}

\item{y}{Optional destinitation object (matrix or big.matrix); if not specified, a big.matrix will be created}

\item{type}{preferably specified (e.g. "integer", "double", etc.)}

\item{separated}{use separated column organization of the data instead of column-major organization;
use with caution if the number of columns is large.}

\item{backingfile}{the root name for the file(s) for the cache of x.}

\item{backingpath}{the path to the directory containing the file backing cache}

\item{descriptorfile}{the name of the file to hold the backingfile description, for subsequent use with
\code{\link[bigmemory]{attach.big.matrix}}; if NULL, the backingfile is used as the root part of the descriptor file name.
The descriptor file is placed in the same directory as the backing files.}

\item{binarydescriptor}{the flag to specify if the binary RDS format should be used for the backingfile description,
for subsequent use with \code{\link[bigmemory]{attach.big.matrix}}; if NULL of FALSE, the dput() file format is used.}

\item{shared}{TRUE by default, and always TRUE if the big.matrix is file-backed. For a non-filebacked big.matrix,
shared=FALSE uses non-shared memory, which can be more stable for large (say, >50% of RAM) objects. Shared memory
allocation can sometimes fail in such cases due to exhausted shared-memory resources in the system.}
}
\value{
a \code{"big.matrix"}
}
\description{
This was written to provide subtraction capabilities to big.matrix objects.  The \code{"daxpy"}
function only works with equal size matrices.  To avoid making another large matrix, this a wrapper of C++ around the
big.matrix object to subtract each element from the provided value.  It likely could use further optimization.
It also includes column and row selection in addition to optionally filebacking.
}

