% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tanhBMIP}
\alias{tanhBMIP}
\title{Provide in-place hyperbolic tangent function for class "big.matrix"}
\usage{
tanhBMIP(x)
}
\arguments{
\item{x}{A \code{"big.matrix"} which must be of type "double" to allow decimals}
}
\value{
a \code{"big.matrix"}
}
\description{
This is used to take each element of a \code{"big.matrix"} to its' hyperbolic tangent.
WARNING!!! This replaces the original matrix.
}
\details{
This function takes each element of a \code{"big.matrix"} to its' hyperbolic tangent.
}
\seealso{
\code{\link[bigmemoryExt]{tanhBM}} for separate big.matrix
}

