% Generated by roxygen2 (4.0.2): do not edit by hand
\name{transposeBM}
\alias{transposeBM}
\title{Produce transpose of a "big.matrix"}
\usage{
transposeBM(x, cols = NULL, rows = NULL, y = NULL, type = NULL,
  separated = NULL, backingfile = NULL, backingpath = NULL,
  descriptorfile = NULL, binarydescriptor = FALSE, shared = TRUE)
}
\arguments{
\item{x}{A \code{"big.matrix"}}

\item{cols}{Possible subset of columns for the transpose; could be numeric, named, or logical}

\item{rows}{Possible subset of rows for the transpose; could be numeric, named, or logical}

\item{y}{Optional destinitation object (matrix or big.matrix); if not specified, a big.matrix will be created}

\item{type}{preferably specified (e.g. "integer", "double", etc.)}

\item{separated}{use separated column organization of the data instead of column-major organization;
use with caution if the number of columns is large.}

\item{backingfile}{the root name for the file(s) for the cache of x.}

\item{backingpath}{}

\item{descriptorfile}{}

\item{binarydescriptor}{}

\item{shared}{}
}
\value{
a transposed \code{"big.matrix"}
}
\description{
This is used to take the transpose of a big.matrix, with the new copy optionally filebacked
}
\details{
This is need to make a tranpose of a \code{"big.matrix"} because R syntax won't access beyond the pointer.
Converting between normal 'matrix' and 'big.matrix' also is expensive to use the normal \code{"t"} function.  This
also allows the user to optionally fileback the transposed object if it will be accessed several times.
}

