% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logBMIP}
\alias{logBMIP}
\title{Provide in-place natural log function for class "big.matrix"}
\usage{
logBMIP(x)
}
\arguments{
\item{x}{A \code{"big.matrix"} which must be of type "double" to allow decimals}
}
\value{
a \code{"big.matrix"}
}
\description{
This is used to take each element of a \code{"big.matrix"} to a specified power.
WARNING!!! This replaces the original matrix.
}
\details{
This function takes each element of a \code{"big.matrix"} to the specified power.
This function should only need to be used if the new matrix must be filebacked as the
NULL call is applied in the \code{\link[base]{log}} method.
}

