% Generated by roxygen2 (4.0.2): do not edit by hand
\name{powBMIP}
\alias{powBMIP}
\title{Provide in-place power (i.e. '^') for class "big.matrix"}
\usage{
powBMIP(x, value)
}
\arguments{
\item{x}{A \code{"big.matrix"}}

\item{value}{A \code{"numeric"} value to take each element to the power}
}
\value{
a \code{"big.matrix"}
}
\description{
This is used to take each element of a \code{"big.matrix"} to a specified power.
WARNING!!! This replaces the original matrix.
}
\details{
This function takes each element of a \code{"big.matrix"} to the specified power.
}
\seealso{
\code{\link[base]{^}} for new big.matrix object \code{\link[bigmemoryExt]{powBM}} for more flexible new big.matrix
}

